MA-18799-2 several coverity issues of opencl
Fix below CID:
12238230 Unsigned compared against 0
12238231 Unsigned compared against 0

Reviewed-on: http://androidsource.ap.freescale.net/project/15281

+++ opencl-2d/opencl-2d.cpp
@@ -292,7 +292,7 @@ static cl_program CreateProgramFromBinary(cl_context context, cl_device_id devic
     size_t binarySize;
     fseek(fp, 0, SEEK_END);
     binarySize = ftell(fp);
-    if(binarySize < 0) {
+    if((int)binarySize < 0) {
         g2d_printf("Error ftell %s.", fileName);
         fclose(fp);
         return NULL;
@@ -357,7 +357,7 @@ static cl_program CreateProgram(cl_context context, cl_device_id device,
     // get the length of the source code
     fseek(pFileStream, 0, SEEK_END);
     program_length = (size_t)ftell(pFileStream);
-    if(program_length < 0)
+    if((int)program_length < 0)
     {
         g2d_printf("Failed to ftell file %s\n", fileName);
         fclose(pFileStream);
